clean up exposed crate features
- Put `TokioExecutor` and `TokioIo` into a single `rt::tokio` module.
- The `rt::tokio` module is behind a `tokio` crate feature.
- Changed `auto` crate feature to `server-auto`.
- Added `client-legacy` crate feature, with `client::legacy` behind it.
- Removed `tcp` and `runtime` features, code now depends on the `tokio`
  feature.
