+++ Cargo.toml
@@ -22,9 +22,9 @@ http = "1.0"
 http-body = "1.0.0"
 bytes = "1"
 pin-project-lite = "0.2.4"
-socket2 = "0.5"
+socket2 = { version = "0.5", optional = true }
 tracing = { version = "0.1", default-features = false, features = ["std"] }
-tokio = { version = "1", features = ["net", "rt", "time"] }
+tokio = { version = "1", optional = true, features = ["net", "rt", "time"] }
 tower-service = "0.3"
 tower = { version = "0.4", features = ["make", "util"] }
 
@@ -42,21 +42,30 @@ pnet_datalink = "0.34.0"
 default = []
 
 # Shorthand to enable everything
-full = ["client", "server", "http1", "http2", "tcp", "auto", "runtime"]
+full = [
+    "client",
+    "client-legacy",
+    "server",
+    "server-auto",
+    "http1",
+    "http2",
+    "tokio",
+]
 
 client = ["hyper/client"]
+client-legacy = ["client"]
+
 server = ["hyper/server"]
+server-auto = ["hyper/server", "http1", "http2"]
 
 http1 = ["hyper/http1"]
 http2 = ["hyper/http2"]
 
-tcp = []
-auto = ["hyper/server", "http1", "http2"]
-runtime = []
+tokio = ["dep:tokio", "dep:socket2"]
 
 # internal features used in CI
 __internal_happy_eyeballs_tests = []
 
 [[example]]
 name = "client"
-required-features = ["client", "http1", "tcp", "runtime"]
+required-features = ["client-legacy", "http1", "tokio"]
