+++ src/client/mod.rs
@@ -1,6 +1,7 @@
 //! HTTP client utilities
 
 /// Legacy implementations of `connect` module and `Client`
+#[cfg(feature = "client-legacy")]
 pub mod legacy;
 #[doc(hidden)]
 pub mod pool;
