+++ src/rt/tokio.rs
@@ -1,11 +1,18 @@
 #![allow(dead_code)]
 //! Tokio IO integration for hyper
 use std::{
+    future::Future,
     pin::Pin,
     task::{Context, Poll},
 };
 
+use hyper::rt::Executor;
 use pin_project_lite::pin_project;
+
+/// Future executor that utilises `tokio` threads.
+#[non_exhaustive]
+#[derive(Default, Debug, Clone)]
+pub struct TokioExecutor {}
 
 pin_project! {
     /// A wrapping implementing hyper IO traits for a type that
@@ -16,6 +23,27 @@ pin_project! {
         inner: T,
     }
 }
+
+// ===== impl TokioExecutor =====
+
+impl<Fut> Executor<Fut> for TokioExecutor
+where
+    Fut: Future + Send + 'static,
+    Fut::Output: Send + 'static,
+{
+    fn execute(&self, fut: Fut) {
+        tokio::spawn(fut);
+    }
+}
+
+impl TokioExecutor {
+    /// Create new executor that relies on [`tokio::spawn`] to execute futures.
+    pub fn new() -> Self {
+        Self {}
+    }
+}
+
+// ==== impl TokioIo =====
 
 impl<T> TokioIo<T> {
     /// Wrap a type implementing Tokio's IO traits.
@@ -159,5 +187,23 @@ where
         bufs: &[std::io::IoSlice<'_>],
     ) -> Poll<Result<usize, std::io::Error>> {
         hyper::rt::Write::poll_write_vectored(self.project().inner, cx, bufs)
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use crate::rt::tokio_executor::TokioExecutor;
+    use hyper::rt::Executor;
+    use tokio::sync::oneshot;
+
+    #[cfg(not(miri))]
+    #[tokio::test]
+    async fn simple_execute() -> Result<(), Box<dyn std::error::Error>> {
+        let (tx, rx) = oneshot::channel();
+        let executor = TokioExecutor::new();
+        executor.execute(async move {
+            tx.send(()).unwrap();
+        });
+        rx.await.map_err(Into::into)
     }
 }
